check for .sbm file extension on --mod
According to the source code, the string ModId includes the .sbm file extension ((publishedFileId.ToString() + ".sbm"));), tested in-game as well. Also, ModName is practically useless to try and parse, since any space in the parameter is not included since we're only taking the first word ("--mod=".Length) and most mod names have spaces in them, and the help context on line 134 says to use only the ID anyway.